Trait crest::request::Request

pub trait Request<'a> {
    fn parameters<I, K, V>(&mut self, params: I) where Self: Sized, I: IntoIterator, I::Item: Borrow<(K, V)>, K: AsRef<str>, V: AsRef<str> { ... }
    fn headers(&mut self) -> &mut Headers { ... }
    fn send(self) -> Result<Response> where Self: Sized { ... }
    fn send_and_into<T>(self) -> Result<T> where Self: Sized, T: Deserialize { ... }
}

Affords core request functionality.

Provided Methods

fn parameters<I, K, V>(&mut self, params: I) where Self: Sized, I: IntoIterator, I::Item: Borrow<(K, V)>, K: AsRef<str>, V: AsRef<str>

Sets the parameters of the request.

If the request is of type POST, the parameters will go in its body, and one header will be set to Content-Type: application/x-www-form-urlencoded.

In any other case, the parameters will go in the URL query string, and no header will be set.

Example:

// assuming a declared `request`
request.parameters(&[
    ("param1", "value1"),
    ("param2", "value2"),
]);

fn headers(&mut self) -> &mut Headers

Gives a mutable reference to the Headers inside a Request.

For example, to declare a header with Connection: close:

use hyper::header;

// assuming a declared `request`
request.headers().set(header::Connection::close());

fn send(self) -> Result<Response> where Self: Sized

Performs the request.

fn send_and_into<T>(self) -> Result<T> where Self: Sized, T: Deserialize

Convenience function to perform a request, deserializing its response.